+++ src/app/(fullscreen)/event-create/components/cinema-skeleton-card.tsx
