+++ src/app/(fullscreen)/event-create/components/step6-place.tsx
@@ -7,7 +7,7 @@ import { useFormContext, useWatch } from "react-hook-form";
 import { LocationIcon } from "@/icons/index";
 import { formatPrice } from "@/utils/format-price";
 import { useSelectedCinemaStore } from "app/_stores/use-selected-cinema-store";
-import CinemaSkeletonCard from "../_components/cinema-skeleton-card";
+import CinemaSkeletonCard from "./cinema-skeleton-card";
 export default function Step6() {
   const { setValue, control } = useFormContext();
   const selectedLocationId = useWatch({ control, name: "locationId" });
